RAC-402 feat : 탈퇴시 소셜 연결 해제
