+++ src/main/java/com/postgraduate/domain/auth/application/usecase/oauth/kakao/KakaoSignOutUseCase.java
@@ -1,10 +1,19 @@
 package com.postgraduate.domain.auth.application.usecase.oauth.kakao;
 
+import com.postgraduate.domain.auth.application.dto.req.SignOutRequest;
 import com.postgraduate.domain.auth.application.usecase.oauth.SignOutUseCase;
 import com.postgraduate.domain.auth.exception.KakaoException;
+import com.postgraduate.domain.senior.domain.entity.Senior;
+import com.postgraduate.domain.senior.domain.service.SeniorGetService;
+import com.postgraduate.domain.user.quit.application.mapper.QuitMapper;
+import com.postgraduate.domain.user.quit.application.utils.QuitUtils;
+import com.postgraduate.domain.user.quit.domain.entity.Quit;
+import com.postgraduate.domain.user.quit.domain.service.QuitSaveService;
 import com.postgraduate.domain.user.user.domain.entity.User;
+import com.postgraduate.domain.user.user.domain.entity.constant.Role;
 import com.postgraduate.domain.user.user.domain.service.UserGetService;
 import com.postgraduate.domain.user.user.domain.service.UserUpdateService;
+import com.postgraduate.domain.user.user.exception.DeletedUserException;
 import com.postgraduate.global.config.security.jwt.util.JwtUtils;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
@@ -24,6 +33,9 @@ public class KakaoSignOutUseCase implements SignOutUseCase {
     private final WebClient webClient;
     private final UserUpdateService userUpdateService;
     private final UserGetService userGetService;
+    private final QuitSaveService quitSaveService;
+    private final SeniorGetService seniorGetService;
+    private final QuitUtils quitUtils;
     private final JwtUtils jwtUtils;
 
     @Value("${admin-id.kakao}")
@@ -32,9 +44,8 @@ public class KakaoSignOutUseCase implements SignOutUseCase {
     private static final String KAKAO_UNLINK_URI = "https://kapi.kakao.com/v1/user/unlink";
 
     @Override
-    public void signOut(Long userId) {
+    public void signOut(User user, SignOutRequest signOutRequest) {
         try {
-            User user = userGetService.byUserId(userId);
             MultiValueMap<String, String> requestBody = getRequestBody(user.getSocialId());
             webClient.post()
                     .uri(KAKAO_UNLINK_URI)
@@ -44,11 +55,30 @@ public class KakaoSignOutUseCase implements SignOutUseCase {
                     .retrieve()
                     .bodyToMono(String.class)
                     .block();
-            userUpdateService.updateDelete(user);
-            jwtUtils.makeExpired(userId);
+            updateDelete(user, signOutRequest);
+            jwtUtils.makeExpired(user.getUserId());
         } catch (WebClientResponseException ex) {
             throw new KakaoException();
         }
+    }
+
+    private void updateDelete(User user, SignOutRequest signOutRequest) {
+        user = userGetService.byUserId(user.getUserId());
+        checkDeleteCondition(user);
+        Quit quit = QuitMapper.mapToQuit(user, signOutRequest);
+        quitSaveService.save(quit);
+        userUpdateService.updateDelete(user);
+    }
+
+    private void checkDeleteCondition(User user) {
+        if (user.isDelete())
+            throw new DeletedUserException();
+        if (user.getRole().equals(Role.SENIOR)) {
+            Senior senior = seniorGetService.byUser(user);
+            quitUtils.checkDeleteCondition(senior);
+            return;
+        }
+        quitUtils.checkDeleteCondition(user);
     }
 
     private MultiValueMap<String, String> getRequestBody(Long socialId) {
