+++ src/main/java/com/postgraduate/domain/auth/presentation/AuthController.java
@@ -5,11 +5,11 @@ import com.postgraduate.domain.auth.application.dto.res.AuthResponse;
 import com.postgraduate.domain.auth.application.dto.res.AuthUserResponse;
 import com.postgraduate.domain.auth.application.dto.res.JwtTokenResponse;
 import com.postgraduate.domain.auth.application.usecase.oauth.SelectOauth;
+import com.postgraduate.domain.auth.application.usecase.oauth.SignOutUseCase;
 import com.postgraduate.domain.auth.application.usecase.oauth.SignUpUseCase;
 import com.postgraduate.domain.auth.application.usecase.jwt.JwtUseCase;
 import com.postgraduate.domain.auth.application.usecase.oauth.SignInUseCase;
 import com.postgraduate.domain.auth.presentation.constant.Provider;
-import com.postgraduate.domain.user.quit.application.usecase.QuitManageUseCase;
 import com.postgraduate.domain.user.user.domain.entity.User;
 import com.postgraduate.global.dto.ResponseDto;
 import io.swagger.v3.oas.annotations.Operation;
@@ -34,7 +34,6 @@ import static com.postgraduate.global.dto.ResponseDto.create;
 public class AuthController {
     private final SelectOauth selectOauth;
     private final SignUpUseCase signUpUseCase;
-    private final QuitManageUseCase quitManageUseCase;
     private final JwtUseCase jwtUseCase;
 
     @PostMapping("/login/token/{provider}")
@@ -95,7 +94,7 @@ public class AuthController {
     @PostMapping("/user/change")
     @Operation(summary = "후배로 추가 가입 | 토큰 필요", description = "대학원생 대학생으로 변경 추가 가입")
     public ResponseEntity<ResponseDto<JwtTokenResponse>> changeUser(@AuthenticationPrincipal User user,
-                                                    @RequestBody @Valid UserChangeRequest changeRequest) {
+                                                                    @RequestBody @Valid UserChangeRequest changeRequest) {
         signUpUseCase.changeUser(user, changeRequest);
         JwtTokenResponse jwtToken = jwtUseCase.changeUser(user);
         return ResponseEntity.ok(create(AUTH_CREATE.getCode(), SUCCESS_AUTH.getMessage(), jwtToken));
@@ -112,7 +111,7 @@ public class AuthController {
     @PostMapping("/senior/change")
     @Operation(summary = "선배로 추가 가입 | 토큰 필요", description = "대학생 대학원생으로 변경 추가 가입")
     public ResponseEntity<ResponseDto<JwtTokenResponse>> changeSenior(@AuthenticationPrincipal User user,
-                                                      @RequestBody @Valid SeniorChangeRequest changeRequest) {
+                                                                      @RequestBody @Valid SeniorChangeRequest changeRequest) {
         User changeUser = signUpUseCase.changeSenior(user, changeRequest);
         JwtTokenResponse jwtToken = jwtUseCase.changeSenior(changeUser);
         return ResponseEntity.ok(create(SENIOR_CREATE.getCode(), CREATE_SENIOR.getMessage(), jwtToken));
@@ -132,11 +131,11 @@ public class AuthController {
         return ResponseEntity.ok(create(AUTH_UPDATE.getCode(), SUCCESS_REGENERATE_TOKEN.getMessage(), jwtToken));
     }
 
-    @PostMapping("/signout")
+    @PostMapping("/signout/{provider}")
     @Operation(summary = "회원 탈퇴", description = "회원 탈퇴 진행")
-    public ResponseEntity<ResponseDto<Void>> signOut(@AuthenticationPrincipal User user, @RequestBody SignOutRequest signOutRequest) {
-        quitManageUseCase.updateDelete(user, signOutRequest);
+    public ResponseEntity<ResponseDto<Void>> signOut(@AuthenticationPrincipal User user, @RequestBody SignOutRequest signOutRequest, @PathVariable Provider provider) {
+        SignOutUseCase signOutUseCase = selectOauth.selectSignOut(provider);
+        signOutUseCase.signOut(user, signOutRequest);
         return ResponseEntity.ok(create(AUTH_DELETE.getCode(), SIGNOUT_USER.getMessage()));
     }
-
 }
